pyproject.toml
[project.scripts]
transcribe-me = "transcribe_me.main:main"

[project]
name = "transcribe-me"
dynamic = ["version"]
authors = [{ name = "echoHello", email = "bash@echohello.dev" }]
description = "A CLI tool to transcribe audio files using OpenAI API"
readme = "README.md"
requires-python = ">=3.12"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "pydub",
    "tqdm",
    "pyyaml",
    "colorama",
    "halo",
    "yamale",
    "tenacity",
    "openai",
    "anthropic",
]

[project.urls]
"Homepage" = "https://github.com/echohello-dev/transcribe-me"
"Bug Tracker" = "https://github.com/echohello-dev/transcribe-me/issues"

[tool.setuptools]
packages = [
    "transcribe_me",
    "transcribe_me.config",
    "transcribe_me.audio",
    "transcribe_me.summarization",
]

[tool.setuptools.package-data]
"transcribe_me.config" = ["*.yaml"]

[tool.setuptools.dynamic]
version = { attr = "transcribe_me.__version__" }
